pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "els-pipeline"
version = "0.1.0"
description = "ELS Normalization Pipeline for Early Learning Standards"
requires-python = ">=3.9"
dependencies = [
    "boto3>=1.28.0",
    "pydantic>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.82.0",
    "moto>=4.1.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.hypothesis]
max_examples = 10
deadline = 1000
